# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "VTEI Student API",
        "description": "REST API для мобільного додатку студентів ВТЕІ",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/schedule/teachers": {
            "get": {
                "tags": [
                    "Schedule"
                ],
                "summary": "Список викладачів згрупованих по кафедрах",
                "description": "Повертає викладачів з розкладу поточної групи, згрупованих по кафедрах.",
                "operationId": "a7dc791515a132df84bba1048ed7dc4d",
                "responses": {
                    "200": {
                        "description": "Масив кафедр з викладачами"
                    },
                    "401": {
                        "description": "Не авторизований"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/schedule/auditoriums": {
            "get": {
                "tags": [
                    "Schedule"
                ],
                "summary": "Список аудиторій згрупованих по корпусах",
                "description": "Повертає аудиторії з розкладу поточної групи, згруповані по корпусах.",
                "operationId": "959e1fad24a4923f59fcb9f0799723c0",
                "responses": {
                    "200": {
                        "description": "Масив корпусів з аудиторіями"
                    },
                    "401": {
                        "description": "Не авторизований"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/schedule/subjects": {
            "get": {
                "tags": [
                    "Schedule"
                ],
                "summary": "Список дисциплін",
                "description": "Повертає дисципліни з розкладу поточної групи.",
                "operationId": "1cfe313cfcc1f98958c5870439c97add",
                "responses": {
                    "200": {
                        "description": "Список дисциплін"
                    },
                    "401": {
                        "description": "Не авторизований"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/schedule/groups": {
            "get": {
                "tags": [
                    "Schedule"
                ],
                "summary": "Список груп згрупованих по факультетах",
                "description": "Повертає всі групи інституту, згруповані по факультетах і курсах.",
                "operationId": "518547062fe0b1e3a57d980e6ef92b7b",
                "responses": {
                    "200": {
                        "description": "Масив факультетів з групами"
                    },
                    "401": {
                        "description": "Не авторизований"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/schedule/lesson-types": {
            "get": {
                "tags": [
                    "Schedule"
                ],
                "summary": "Список типів занять",
                "description": "Повертає всі можливі типи занять з ключами для фільтрації.",
                "operationId": "748e003886703555b7e56a34cda1b684",
                "responses": {
                    "200": {
                        "description": "Список типів занять"
                    },
                    "401": {
                        "description": "Не авторизований"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
        "/settings": {
            "get": {
                "tags": [
                    "Settings"
                ],
                "summary": "Налаштування студента",
                "description": "Повертає поточні налаштування: FCM токен, Telegram.",
                "operationId": "2e06d6f946c7ac2bea504994bae0224d",
                "responses": {
                    "200": {
                        "description": "Налаштування"
                    },
                    "401": {
                        "description": "Не авторизований"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/settings/notifications": {
            "put": {
                "tags": [
                    "Settings"
                ],
                "summary": "Оновити FCM токен",
                "description": "Зберігає або оновлює FCM токен для push-нотифікацій. Передати null щоб відключити.",
                "operationId": "3bc500a2dc48b8e709fc09ab919dd96d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "fcm_token": {
                                        "description": "FCM токен пристрою або null",
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Оновлено"
                    },
                    "401": {
                        "description": "Не авторизований"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}